[WEB-3322] - moved types
